Remove `sanitize-html` dependency
Closes #18606

The `sanitize-html` dependency is pulled in just to sanitize the
HTML content returned from calls to the WordPress.org API for
plugin descriptions and other data. It's 160kB in total and
about 50kB gzip'd.

Instead of pushing it into a vendor bundle or asynchronously
loading it I think that we can simply eliminate it.

Our goal is to sanitize HTML and prevent cross-site scripting
attacks (XSS) but we should be able to do this easily enough
by harnesing the browser's built-in HTML parser and walking
through each node to clean them up. Anything that's not
explicitly allowed will either be removed or escaped to prevent
any possibility of running JavaScript.

Since this PR introduces a custom parser I took the libery of
adding in support for a few more tags (`div` and `span`) as
well as support for embedding YouTube videos, figuring that if
we are conservative enough we won't have to worry about trouble
including those.

Obviously this needs some tests, even some rudimentary tests.
Also it should be cleaned up. For now, I'm hoping this patch
gathers enough feedback to see if we support this approach.

For now, I'm seeing that the new code is about 1700B and only
about 850B gzip'd, which is significantly smaller than it was
with the dependency.